Refactored to simplify
Also added an (empty) broadcast receiver for package changes.

How to show the rest of the apps? If I use an icon to show a
sub menu, then the finger left the screen and the menu would
be open.

Then there's a configuration activity required to configure
what's in the pie menu and what not.