Show all apps only when Home is pressed at rest
Remember when the activity was paused to tell if the Home button
was pressed immediately before getting onNewIntent().

Otherwise a user could press the Overview button (what triggers
onPause() too) and *then* the Home button. In which case we don't
want to show all apps but just the blank home screen.