Replace SurfaceView with a simple View
A SurfaceView is more complex and should only be used when
rendering from a separate thread. Since the PieMenu is rendered
on the main thread, there is no advantage in using a SurfaceView.

Using a simple View actually improves performance on some devices
(e.g. my Fairphone 2) while it remains to be the same on others.