Register broadcast listener in application
It's important to register the receiver in the application context
to bind the receiver to the application object. Using an Activity
context silently stops working after while before Android O.

Note that it's not required to unregister the receiver because it
needs to be there as long as this application is running.