Start a new fling when the view height changes
OverScroller.notifyVerticalEdgeReached() animates until finalX
what means the list is scrolling all by itself when the view
height changes. Not what we want.

OverScroller.springBack() stops the current fling so the list
suddendly stops scrolling once the view changes. Also not what
we want.

So the only option left is to stop the current fing and simply
start a new one instead.