Request to ignore battery optimizations
And explain why this is a good idea in an extra Activity.

Disable battery optimizations because the whole point of a home
screen launcher is to be available all the time. This app does not
consume battery when not in use. And it should always and immediately
be available (and thus in memory).

Killing this process in Doze mode and reloading it when the user
unlocks their device actually consumes more battery.