Remove static locale and simply call getDefault()
This might be a little slower because Locale.getDefault() is
called over and over again but it's still better than storing
a Locale in a static final field because this won't change when
the user is changing the Locale while the app is running.

Of course, it'd be best to listen for configuration changes and
update a local field from Locale.getDefault() but that's a lot
more code for an indistinguishable difference.