Refactor BatteryOptimization to SettingsActivity
And put the preferences into the new SettingsActivity.

Show a short welcome message instead of the long explanation
about battery optimization. This way the option to set the
default launcher becomes equally prominent, which is good
because I see a lot of reviews complaining that the launcher
would not stick or the home gesture isn't working or things
like that.