Convert search strictness enum to integers
While an enum is the textbook choice for this kind of situation,
simple integer constants are just a better fit here (in my opinion).

On Android, many options are just integer constants (like
`ActivityInfo.SCREEN_ORIENTATION_*` which we also use here),
and this way we can avoid the String parsing from `Enum.valueOf`
when restoring the enum from a `SharedPreference`.