Revert option for an edit menu button
The reason for this alternative was to "lock" the menu because
it was too easy to enter the menu editor, which was because of
a bug handling the scroll reference (see 499d23), I believe.

Now that this is resolved, there is no reason for having that
extra edit button anymore. Also, disabling long press in the
app drawer means that apps can no longer be uninstalled easily,
because that also works by grabbing an app icon from the app
drawer.

This reverts commit 6b43e5ecad6311663de5704a2b290ced34eb81ff.