Improve and shorten tips
Shorten some tips to make them fit into one line. Hopefully.

Mention that you can circle and icon around the pie to quickly
change the icon order. Probably too subtle to really be noticed,
but it's a start.

I would like to keep the tips short and sweet as longer texts
are probably simply overlooked and not read. But this isn't too
easy now that there are some translations. So maybe I will need
to add support for multi-line tips.