Use space to move selection for IME_ACTION_SEARCH
Use trailing space(s) in the search query to move the app selector
for the IME_ACTION_SEARCH action on the soft keyboard. This way it's
quick and easy to select the next best result and hit "enter" to
launch the app.

Also a setting was added to choose and keep the previous double-space-
launches-app behaviour instead.