Fix tap detection
A tap does not time out after `ViewConfiguration.getTapTimeout()`,
which specifies the "duration […] we will wait to see if a touch
event is a tap or a scroll".

Rather, any touch is a tap until it becomes a long press.

Previously, all taps that lasted longer than `tapTimeout` were
incorrectly discarded.