Revert option to blur window background
Needs more testing.

Before Android 12, when `windowIsTranslucent` is `true`, you can
see the Activity below the current one even if it belongs to the
same app.

From Android 12 on, you only see Activities from *other* apps, but
not your own. Now this is the behaviour we need in this app, because
when blur is not available, our own Activity below should *not* be
visible, of course.

Unfortunately, I can't find any documentation about this change.
So I can't be sure if this is indeed a change, or just happens to
be so on my own devices.

This reverts commit 3c2c3ed7d4d4922405cf1a6559f8e908853c729a.