Fix storing/restoring icon mappings
I am an idiot. Object serialization can't work (over app updates)
when the code is obfuscated, of course.

Object serialization works with the *name* of an object, and R8
shortens the names to some abbreviation that can and will vary
from build to build.

So yes, this means everyone will loose all icon mappings with the
next update. Sorry.