Support changing pie geometry
Now the first icon can be at 0, 90, 180 or 270 degrees to make it
possible to have an app in the bottom center of the pie when 6 apps
are used (for example).

Also, the icons can now be scaled down in their cell, which may be
desirable when only few icons are in the pie.

The scale listener (and gesture) is now also replaced with a button.
The scale gesture was always a bit cumbersome.