Add an option to always open app list on a tap
Even if the tap takes longer than the long press timeout.

The advantage is that any touch (on the home screen) has an effect.
Either the tap is launching an app or it opens the app listing.

Before, any touch that took longer than the long press timeout (and
that didn't select an app to launch) did nothing, and the pie menu
just vanished.