Fix modes when returning after icon change
Because `fadeOutMode()` wasn't called when `keepMode` was `true`
in `hideList()`.

Since `keepMode` has always been an ugly hack and another complication,
I have now removed it completely. The only benefit of it was to return
to the editor after changing an icon from it and that is a rather rare
use case that doesn't justify this complication.