Fix Android 14 gesture navigation recent apps bug
Or rather make it a lot less frequently happening.

The bug that seems to be introduced with Android 14 can be reproduced
with the following steps:

1. only for gesture navigation
2. make the home gesture twice (swipe from the bottom of the screen up)
3. enter list of recent apps
4. now an empty, broken screen (activity?) is in the list of recent apps

This broken screen can be selected, but doesn't activate the home
activity but just shows nothing (the wallpaper shines through).

I have also tried this with other popular launchers (e.g. Niagara Launcher,
Nova Launcher etc.) and they have all shown the same error.