Add option to force relaunch home activity
To prevent Android from restarting an existing activity.
Which seems to be a problem on some fold devices, see issue #127.

I have not been able to reproduce this issue, not even on a foldable
device, and this is most likely not the best way to fix the issue as
activity reuse is an integral part of Android. Still, I need a solution
for the people who have this problem.