Make forcing HomeActivity restart the default
Since the invisible drawer bug is happening to more and more users.
See: https://github.com/markusfisch/PieLauncher/issues/149

Unfortunately, the reason for this bug is still unclear.

It occurs very rarely for me, almost never, and it's really difficult
to get information about the bug when it occurs. I still haven't found
a way to reliably reproduce the error.

But since users are hit by this, I will enable this option by default.