Revert to postpone app indexing when editing pie
Doesn't work when icons need to be updated (e.g. when changing the
icon set) or sometimes even on uninstalls.

This approach is too complex, and it's better to revert to the code
before. Even when updates while editing the pie menu are still very
much annoying. Let's just find a better way.

This reverts commit edd4ee6f5d86bbc47be0182d137b8f484cb8f295.